+++ src/aibase.C
@@ -168,7 +168,7 @@ char nbuff[16];
 }
 
 // Push a maneuver on to the stack
-sManeuverState *aiBase::PushManeuver(int mnvr, unsigned long flags, sREAL d0,
+sManeuverState *aiBase::PushManeuver(int mnvr, uint32_t flags, sREAL d0,
 												sREAL d1, sREAL d2)
 {
 	if (mnvrStackPtr < aiB_MNVR_MAX - 1)
@@ -187,7 +187,7 @@ sManeuverState *aiBase::PushManeuver(int mnvr, unsigned long flags, sREAL d0,
 		return (NULL);
 }
 
-sManeuverState *aiBase::PushManeuverX(int mnvr, unsigned long flags,
+sManeuverState *aiBase::PushManeuverX(int mnvr, uint32_t flags,
 													sREAL d0, sREAL d1, sREAL d2,
 													sREAL d3, sREAL d4, sREAL d5)
 {
@@ -242,7 +242,7 @@ void aiBase::ClearManeuverStack()
 
 // Set a maneuver. This clears the stack
 // and makes the maneuver the current one
-void aiBase::SetManeuver(int mnvr, unsigned long flags, 
+void aiBase::SetManeuver(int mnvr, uint32_t flags, 
 									sREAL d0, sREAL d1, sREAL d2)
 {
 	if (mnvr >= 1 && mnvr < sManeuver::LAST)
@@ -257,7 +257,7 @@ void aiBase::SetManeuver(int mnvr, unsigned long flags,
 			mnvr);
 }
 
-void aiBase::SetManeuverX(int mnvr, unsigned long flags, 
+void aiBase::SetManeuverX(int mnvr, uint32_t flags, 
 									sREAL d0, sREAL d1, sREAL d2,
 									sREAL d3, sREAL d4, sREAL d5)
 {
@@ -323,12 +323,12 @@ void __cdecl aiBase::LOG(const char *format, ...)
 /**********************************************************
  * Attack list routines                                   *
  **********************************************************/
-void aiBase::AddAttacker(unsigned long attackerIdx)
+void aiBase::AddAttacker(uint32_t attackerIdx)
 {
 	attackList.AddAttacker(attackerIdx);
 }
 
-void aiBase::DeleteAttacker(unsigned long attackerIdx)
+void aiBase::DeleteAttacker(uint32_t attackerIdx)
 {
 	attackList.DeleteAttacker(attackerIdx);
 }
